Made CoinSprites use the BlockContainer SpriteContainer.
The coins and blocks are all on the same sprite sheet. There's no need to create a CoinContainer SpriteContainer that is an exact duplicate of BlockContainer.

By reusing BlockContainer, all of the coin game objects are sharing the same material as the blocks. This forces dynamic draw call batching, and reduces the number of draw calls per frame.